Source files and tests for Kubernetes frontend
Adding source/tests for k8s frontend. This includes main.go (program startup), orchestration interface,
and k8s implementation of the interface. Also adding the controller package which will act as a middle layer
between the orchestration and ipam system. The controller will use a store for data (not yet used).

The Kubernetes portion watches Ingresses/ConfigMaps for an ipam annotation. If found, IPGroups are populated, which
sort resources and their hosts by a common desired IP address. This data is then sent on to the controller.