Infoblox support and controller integration
This commit adds the manager interface with the infoblox client, which talks to infoblox and can create and delete records. It can also
get and reserve IP addresses, as well as release them. These additions also integrate the controller middle layer to communicate
with the infoblox client and use the ipGroups from the orchestration to create/update/delete records on infoblox. The controller
keeps an internal store of IP to host mappings, in order to keep the state between the orchestration and the ip manager consistent.

Also fixed an issue in the kubernetes piece that prevented empty configs from being written out on startup (when necessary).